+++ src/commonMain/kotlin/com.bay.aiclient/api/google/Internal.kt
@@ -26,7 +26,7 @@ data class GoogleHttpGenerationConfig(
     val maxOutputTokens: Int? = null,
     val temperature: Double? = null,
     val topP: Double? = null,
-    val topK: Double? = null,
+    val topK: Int? = null,
     val presencePenalty: Double? = null,
     val frequencyPenalty: Double? = null,
 )
